requirements.txt
flask==2.2.3
waitress==2.1.2
supabase==1.0.3
google-generativeai==0.3.1
python-dotenv==1.0.0
pytz==2023.3
